requirements.txt
botcity-framework-core
requests
pygetwindow
